Add new request building API
Oh boy this is a big one. The protocol previously known as `Request` is now
called `RequestConvertible`. This better reflects[^1] what it does and means we
free up `Request` for a new type.

The new `Request` type is a concrete implementation of `RequestConvertible`. It
provides a set of functions that can be chained together to customise a
request. The request type works in tandem with the new `RequestProviding`
type. The `RequestProviding` type provides the base URL for a request and a
`Request` is generic over a `RequestProviding` type.

The intention with `RequestProviding` is that users will create a new
`RequestProviding` type for each API they work with. The type provides the
initial base request for the API with some pre-configured parameters that are
common to that API's requests. Additionally, extensions on the `Request` type
can be constrained to the API so users can create a mini-DSL for the API's
requests.

[^1]: An even better name would be `URLRequestConvertible` but I believe
Alamofire has protocol with the same name IIRC.