Refactor URLSession perform(_:) extensions
This commit greatly simplifies the implementation of the extensions. Rather than
having two separate implementations for requests with and without response
bodies there's now a single implementation. If the response type is Void then
the body decoding step is skipped.

The implementation of the perform(_:) method has been simplified so that it's
easier to follow. In doing so, I've fixed a bug where client side errors that
occur after the initial response are not passed through to the completion
block. The URL loading system doesn't seem to let this situation happen (if
there's a client side error, the `URLResponse` is always set to nil) but as
there is no type-level guarantee, I've added code to handle this anyway.

The perform(_:) method now accepts an additional `configureTask` block. This
provides an opportunity for users to customise the URLSessionTask and set up
key-value observers before it is resumed.

This commit also introduces extensive test coverage for the perform(_:) method.