Add encoding system for request bodies
The system is similar to the decoding system for response bodies. A new type,
the BodyProvider, has been introduced that handles encoding some data into a
request. Like the ResponseDecoder, the type wraps a closure that implements the
actual logic of creating the response body. The closure is also responsible for
updating the header of the request with fields like the content-type of the
body.

Predefined BodyProviders for raw data, textual bodies and json encoded bodies
have been provided. Additional providers (e.g.,
application/x-www-form-urlencoded, multipart/form-data) will be the focus of
a future release of Requests.