Update URLSession extension to use the standard lib's Result type
This is quite nice. Since the standard library's Result type has a strongly
typed error parameter, I've introduced a new error type---the
RequestTransportError. This type provides a lot more contextual information than
the RequestError type (which is still around). It captures the underlying error,
the request and the response if one was received.

This is an API breaking change but the effect is limited. For the success case,
nothing changes since the standard library's success case matches with
Request's. For the error case, the response is now part of the wrapped error and
all errors are now RequestTransportErrors. The type system will catch this
change so there's no risk of this introducing run time errors.